split up commands from commands.js into multiple files
