+++ workspace sir vivor/commands/authText.js
@@ -5,7 +5,7 @@
  */
 
 module.exports = {
-	
+
 	'2g': '==2guhde4u',
 	'2g4u': '==2guhde4u',
 	'2guhde': '==2guhde4u',
@@ -14,7 +14,7 @@ module.exports = {
 	ahr: '==ahelpfulrayquaza',
 	ray: '==ahelpfulrayquaza',
 	ahelpfulrayquaza: 'Skill issue',
-	
+
 	aud: '==audiino',
 	audi: '==audiino',
 	audii: '==audiino',
@@ -23,39 +23,39 @@ module.exports = {
 	ii: '==audiino',
 	ii88: '==audiino',
 	audiino: 'You don’t need to tell me that, I audiino',
-	
+
 	bea: '==beautifications',
 	beautifications: 'brb changing my username and pfp',
-	
+
 	bon: '==bondance',
 	bondance: 'Lol, more like can\'t dance',
-	
+
 	bread: '==breadloeuf',
 	breadloeuf: 'You\'re toast *explodes u with mind power*',
-	
+
 	myster: '==bugbuzzing',
 	myster17: '==bugbuzzing',
 	bb: '==bugbuzzing',
 	bugbuzzing: 'buzz buzz',
-	
+
 	pak: '==ceteris',
 	cet: '==ceteris',
 	ceterisparibus: '==ceteris',
 	ceteris: 'CHOO CHOO ALL ABOARD THE PARTYBUS TRAIN',
-	
+
 	cheese: 'Muffinss',
-	
+
 	crab: '==crabachillable',
 	crabachillable: '/me crabwalks',
-	
+
 	deetah: 'Survivor\'s Resident Cheetah',
-	
+
 	dl: '==dontlose',
 	dontlose: 'dont lose? HA! More like dont win :^)',
-	
+
 	eoth: '==easyonthehills',
 	easyonthehills: '/me resigns',
-	
+
 	unleashourpassion: '==mez',
 	uop: '==mez',
 	akasianse: '==mez',
@@ -65,104 +65,104 @@ module.exports = {
 
 	gts: '==gary',
 	gary: 'shits rigged',
-	
+
 	guishark: '==gui',
 	gui: 'Hi [target],  I\'m Gui♥Shark!',
-	
+
 	monkey: '==henka',
 	henka: 'H E N K A B O Y S',
-	
+
 	inac: '==inactive',
 	inactive: 'NOW',
-	
+
 	ledepression: '==depress',
 	depress: 'if this had the golf mod it\'d be a different story...',
-	
+
 	lfo: '==lfolfo',
 	lfolfo: 'It\’s loaf-o loaf-o thank you very much',
 
 	ljb: 'hi frens :3',
-	
+
 	lunarixis: '==lunar',
 	lunar: '/me sips tea',
-	
+
 	opple: '==flopple',
 	flopple: 'Who\'s the floppling? Opple? AGAIN?',
-	
+
 	micro: '==micromorphic',
 	micromorphic: 'Hello everyone, unfortunately I have some very sudden news.',
-	
+
 	mit: '==mitsuki',
 	mitsuki: 'isso significa algo em português',
-	
+
 	neon: '==neonalphaa',
 	neonalphaa: 'i\'m going to bully someone',
-	
+
 	omroom: '==om',
 	om: '"OM Goodness! Leave OM Room alone..." ;~;',
 
 	marill: '==pm',
 	pacificmarill: '==pm',
 	pm: 'My Huge Power is my terrible rolls... ᘛ⁐̤ᕐᐷ',
-	
+
 	pq: '==penquin',
 	ppq: '==penquin',
 	penquin: 'S-S-Senpai!',
-	
+
 	ptoad: 'Congratulations to ptoad on getting into Washington University!',
-	
+
 	rain: '==rainshaft',
 	rainshaft: ':P',
-	
+
 	sc: '==saltiestcactus43',
 	salti: 'saltiestcactus43',
 	saltiestcactus43: 'Salti plays connect 5 while the rest play connect 4...',
-	
+
 	shadecession: '==shade',
 	shade: 'Better put on my Shadecessions ⌐■_■',
-	
+
 	sifi: '==sificon',
 	sificon: 'If you tabfail my name for Sir.Vivor - you owe me a dollar :>',
-	
+
 	soccer: '==soccerz12',
 	soccerz12: 'Hey, what\'s up guys it\'s Soccer here',
-	
+
 	spieky: 'Global Roomowner, Administrator and all around Good Guy!',
-	
+
 	swirl: '==swirlyder',
 	swirlyder: 'I swear, he\'s not my real dad',
-	
+
 	vex: '==thevexinator',
 	thevexinator: 'vox',
-	
+
 	tikitik: '==tiki',
 	tiki: 'Survivor\'s Resident Russian Spy',
-	
+
 	tush: '==tushavi',
 	tushavi: 'THEME GOD',
-	
+
 	zygten: '==zyg',
 	zyg: '/me sighs... what is there to say?',
 
 	yoshman: '==yosh',
 	yosh: 'Yoshing mode activated, set to full power!',
-	
+
 	anime: '==weeb',
 	weeb: 'Anime was a mistake',
-	
+
 	dn: '==deeznuts',
 	deeznuts: '[target]? More like deez nuts :^)',
-	
+
 	meow: 'meow!',
-	
+
 	pipeitup: '/ME dabs',
-	
+
 	sips: '==sip',
 	sip: '/me sips [target]\'s tea',
-	
+
 	woof: 'woof!',
-	
+
 	saw: '==workshop',
 	workshop: '<<survivorworkshop>>',
-}
+};
