+++ workspace sir vivor/commands/roomOwner.js
@@ -1,0 +1,80 @@
+	/***********************************************
+	 *             ROOM OWNER COMMANDS             *
+	 ***********************************************/
+module.exports = {
+
+	settings: 'set',
+	set: function (arg, user, room) {
+		if (room === user || !user.hasRank(room.id, '#')) return false;
+
+		var opts = arg.split(',');
+		var cmd = toId(opts[0]);
+		var roomid = room.id;
+		if (cmd === 'm' || cmd === 'mod' || cmd === 'modding') {
+			let modOpt;
+			if (!opts[1] || !CONFIGURABLE_MODERATION_OPTIONS[(modOpt = toId(opts[1]))]) {
+				return this.say(room, 'Incorrect command: correct syntax is ' + Config.commandcharacter + 'set mod, [' +
+					Object.keys(CONFIGURABLE_MODERATION_OPTIONS).join('/') + '](, [on/off])');
+			}
+			if (!opts[2]) return this.say(room, 'Moderation for ' + modOpt + ' in this room is currently ' +
+				(this.settings.modding && this.settings.modding[roomid] && modOpt in this.settings.modding[roomid] ? 'OFF' : 'ON') + '.');
+
+			if (!this.settings.modding) this.settings.modding = {};
+			if (!this.settings.modding[roomid]) this.settings.modding[roomid] = {};
+
+			let setting = toId(opts[2]);
+			if (setting === 'on') {
+				delete this.settings.modding[roomid][modOpt];
+				if (Object.isEmpty(this.settings.modding[roomid])) delete this.settings.modding[roomid];
+				if (Object.isEmpty(this.settings.modding)) delete this.settings.modding;
+			} else if (setting === 'off') {
+				this.settings.modding[roomid][modOpt] = 0;
+			} else {
+				return this.say(room, 'Incorrect command: correct syntax is ' + Config.commandcharacter + 'set mod, [' +
+					Object.keys(CONFIGURABLE_MODERATION_OPTIONS).join('/') + '](, [on/off])');
+			}
+
+			this.writeSettings();
+			return this.say(room, 'Moderation for ' + modOpt + ' in this room is now ' + setting.toUpperCase() + '.');
+		}
+
+		if (!(cmd in Commands)) return this.say(room, Config.commandcharacter + '' + opts[0] + ' is not a valid command.');
+
+		var failsafe = 0;
+		while (true) {
+			if (typeof Commands[cmd] === 'string') {
+				cmd = Commands[cmd];
+			} else if (typeof Commands[cmd] === 'function') {
+				if (cmd in CONFIGURABLE_COMMANDS) break;
+				return this.say(room, 'The settings for ' + Config.commandcharacter + '' + opts[0] + ' cannot be changed.');
+			} else {
+				return this.say(room, 'Something went wrong. PM Morfent or TalkTakesTime here or on Smogon with the command you tried.');
+			}
+
+			if (++failsafe > 5) return this.say(room, 'The command "' + Config.commandcharacter + '' + opts[0] + '" could not be found.');
+		}
+
+		if (!opts[1]) {
+			let msg = '' + Config.commandcharacter + '' + cmd + ' is ';
+			if (!this.settings[cmd] || (!(roomid in this.settings[cmd]))) {
+				msg += 'available for users of rank ' + ((cmd === 'autoban' || cmd === 'banword') ? '#' : Config.defaultrank) + ' and above.';
+			} else if (this.settings[cmd][roomid] in CONFIGURABLE_COMMAND_LEVELS) {
+				msg += 'available for users of rank ' + this.settings[cmd][roomid] + ' and above.';
+			} else {
+				msg += this.settings[cmd][roomid] ? 'available for all users in this room.' : 'not available for use in this room.';
+			}
+
+			return this.say(room, msg);
+		}
+
+		let setting = opts[1].trim();
+		if (!(setting in CONFIGURABLE_COMMAND_LEVELS)) return this.say(room, 'Unknown option: "' + setting + '". Valid settings are: off/disable/false, +, %, @, #, &, ~, on/enable/true.');
+		if (!this.settings[cmd]) this.settings[cmd] = {};
+		this.settings[cmd][roomid] = CONFIGURABLE_COMMAND_LEVELS[setting];
+
+		this.writeSettings();
+		this.say(room, 'The command ' + Config.commandcharacter + '' + cmd + ' is now ' +
+			(CONFIGURABLE_COMMAND_LEVELS[setting] === setting ? ' available for users of rank ' + setting + ' and above.' :
+				(this.settings[cmd][roomid] ? 'available for all users in this room.' : 'unavailable for use in this room.')));
+	}
+};
